requirements.txt
langchain>=0.2.0
langchain-groq>=0.1.0
langchain-core>=0.2.0
langgraph>=0.1.0
sentence-transformers>=3.0.0
numpy>=1.26.0
python-dotenv>=1.0.0
torch>=2.0.0
